modif pour post

+++ src/main/java/fr/epsi/b3devc1/msprapi/repository/RegionRepository.java
@@ -3,5 +3,5 @@ package fr.epsi.b3devc1.msprapi.repository;
 import fr.epsi.b3devc1.msprapi.model.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegionRepository extends JpaRepository<Region, Integer> {
+public interface RegionRepository extends JpaRepository<Region, Long> {  // Changer Integer par Long ici
 }
